+++ src/Block/Adminhtml/CustomOrderFees/Report/Grid.php
@@ -34,6 +34,7 @@ class Grid extends AbstractGrid
         $baseFeeAmount = [];
         $paidFeeAmount = [];
         $paidOrderCurrency = [];
+        $baseInvoicedFeeAmount = [];
         $invoicedFeeAmount = [];
         $baseRefundedFeeAmount = [];
         $refundedFeeAmount = [];
@@ -42,6 +43,7 @@ class Grid extends AbstractGrid
             $baseFeeAmount[] = $item->getData('base_fee_amount');
             $paidFeeAmount[] = $item->getData('paid_fee_amount');
             $paidOrderCurrency[] = $item->getData('paid_order_currency');
+            $baseInvoicedFeeAmount[] = $item->getData('base_invoiced_fee_amount');
             $invoicedFeeAmount[] = $item->getData('invoiced_fee_amount');
             $baseRefundedFeeAmount[] = $item->getData('base_refunded_fee_amount');
             $refundedFeeAmount[] = $item->getData('refunded_fee_amount');
@@ -50,6 +52,7 @@ class Grid extends AbstractGrid
         $totals->setData('base_fee_amount', (string) array_sum($baseFeeAmount));
         $totals->setData('paid_fee_amount', implode(', ', $paidFeeAmount));
         $totals->setData('paid_order_currency', implode(', ', $paidOrderCurrency));
+        $totals->setData('base_invoiced_fee_amount', implode(', ', $baseInvoicedFeeAmount));
         $totals->setData('invoiced_fee_amount', implode(', ', $invoicedFeeAmount));
         $totals->setData('base_refunded_fee_amount', implode(', ', $baseRefundedFeeAmount));
         $totals->setData('refunded_fee_amount', implode(', ', $refundedFeeAmount));
@@ -125,6 +128,23 @@ class Grid extends AbstractGrid
                 'rate' => $rate,
                 'header_css_class' => 'col-paid-fee-amount',
                 'column_css_class' => 'col-paid-fee-amount',
+            ],
+        );
+        $this->addColumn(
+            'base_invoiced_fee_amount',
+            [
+                'header' => __('Invoiced'),
+                'type' => 'currency',
+                'index' => 'base_invoiced_fee_amount',
+                'total' => 'sum',
+                'sortable' => false,
+                'renderer' => Currency::class,
+                'rate' => $rate,
+                'header_css_class' => 'col-base-invoiced-fee-amount',
+                'column_css_class' => 'col-base-invoiced-fee-amount',
+                'visibility_filter' => [
+                    'show_base_invoiced_amount',
+                ],
             ],
         );
         $this->addColumn(